Create ClearNuGetLocalResourcesCommand
